+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-## Unreleased
+## 0.9.2
 
 ### Features
 
 - Tag whether event was created in the UI thread ([#495](https://github.com/getsentry/sentry-unity/pull/495))
+
+### Fixes
+
+- Fix missing Sentry/Sentry.h ([#504](https://github.com/getsentry/sentry-unity/pull/504))
+- Automated Android symbols upload now correctly escapes sentry-cli executable path ([#507](https://github.com/getsentry/sentry-unity/pull/507))
 
 ## 0.9.1